feat(tui): add Shift+E export picker and Address notes AI hub action
Move markdown export into er-engine for TUI/desktop sharing, add a multiselect clipboard export modal, and replace AI hub copy-json actions with an agent-driven Address notes flow.
